requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-json-logger==3.3.0
email-validator==2.2.0
pytest==8.4.1
httpx==0.28.1
